refactor: btc hook
